Avoid using MapShapedWriter for coproduct derivation
