+++ modules/generic/src/main/scala/pureconfig/generic/MapShapedWriter.scala
@@ -45,25 +45,4 @@ object MapShapedWriter {
         }
       }
     }
-
-  implicit def cNilWriter[Wrapped]: MapShapedWriter[Wrapped, CNil] = new MapShapedWriter[Wrapped, CNil] {
-    override def to(t: CNil): ConfigValue =
-      throw new IllegalStateException("Cannot encode CNil. This is likely a bug in PureConfig.")
-  }
-
-  final implicit def cConsWriter[Wrapped, Name <: Symbol, V <: Wrapped, T <: Coproduct](
-    implicit
-    coproductHint: CoproductHint[Wrapped],
-    vName: Witness.Aux[Name],
-    vConfigWriter: Derivation[Lazy[ConfigWriter[V]]],
-    tConfigWriter: Lazy[MapShapedWriter[Wrapped, T]]): MapShapedWriter[Wrapped, FieldType[Name, V] :+: T] =
-    new MapShapedWriter[Wrapped, FieldType[Name, V] :+: T] {
-      override def to(t: FieldType[Name, V] :+: T): ConfigValue = t match {
-        case Inl(l) =>
-          coproductHint.to(vConfigWriter.value.value.to(l), vName.value.name)
-
-        case Inr(r) =>
-          tConfigWriter.value.to(r)
-      }
-    }
 }
